requirements.txt
afew==3.0.2.dev82+g23b5aeaa4
alabaster==1.0.0
argcomplete==3.6.2
asgiref==3.11.1
attrs==25.4.0
autocommand==2.2.2
babel==2.17.0
Brlapi==0.8.7
btrfsutil==6.19
certifi==2026.1.4
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dasbus==1.7
dbus-python==1.4.0
defusedxml==0.7.1
distro==1.9.0
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dkimpy==1.1.8
dnspython==2.8.0
docutils==0.22.3
filelock==3.24.0
idna==3.11
imagesize==1.4.1
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.6
libfdt==1.7.2
louis==3.36.0
lxml==6.0.2
MarkupSafe==3.0.3
meson==1.10.1
more-itertools==10.8.0
netsnmp-python==1.0a1
nftables==0.1
notmuch2==0.40
packaging==26.0
pipx==1.8.0
pkg_resources==81.0.0
platformdirs==4.9.3
psutil==7.2.2
pwquality==1.4.5
pycairo==1.29.0
pycparser==2.23
pygdbmi==0.11.0.0
Pygments==2.19.2
PyGObject==3.54.5
PyJWT==2.12.0
pytz==2025.2
pyxdg==0.28
Reflector==2023.6.28.0.36.1
requests==2.32.5
roman-numerals-py==3.1.0
sentry-sdk==2.54.0
setproctitle==1.3.7
setuptools==82.0.0
smbus==1.1
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.5
systemd-python==235
TBB==0.2
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
userpath==1.9.2
wheel==0.46.3
